requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
fastapi==0.104.1
uvicorn==0.24.0
python-decouple==3.8
psycopg2-binary==2.9.9
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
